seperated message form into own component to prevent rerendering posts component
