+++ client/src/components/TweeterChat.jsx
@@ -1,20 +1,19 @@
 import React, { useState, useEffect } from "react";
 import Post from "./Post";
-import arrow from "../assets/arrow.svg";
+import MessageInput from "./MessageInput";
 
 export default function TweeterChat({ socket }) {
-  const [message, setMessage] = useState("");
   const [posts, setPosts] = useState([]);
   const [room, setRoom] = useState("1");
 
   useEffect(() => {
     const joinRoom = () => {
       if (room !== "") {
-       socket.emit("join_room", room);
+        socket.emit("join_room", room);
       }
     };
     joinRoom();
-  }, [])
+  }, [room, socket]);
 
   useEffect(() => {
     socket.on("receive_message", (data) => {
@@ -27,26 +26,8 @@ export default function TweeterChat({ socket }) {
     };
   }, [socket]);
 
-  const handleMessageChange = (event) => {
-    setMessage(event.target.value);
-  };
-
-  const sendPost = () => {
+  const sendMessage = (message) => {
     socket.emit("send_message", { message, room, user: { name: "Owen Kanzler", avatar: "" } });
-  };
-
-  const handleFormSubmit = (e) => {
-    e.preventDefault();
-    if (message.trim() === "") return;
-
-    const newPost = {
-      user: { name: "Owen Kanzler", avatar: "" },
-      message
-    };
-
-    sendPost();
-    // setPosts((prevPosts) => [...prevPosts, newPost]);
-    setMessage("");
   };
 
   return (
@@ -60,24 +41,7 @@ export default function TweeterChat({ socket }) {
           />
         ))}
       </div>
-      <form
-        onSubmit={handleFormSubmit}
-        className="flex gap-4 bg-interactive py-4 px-6 rounded-lg border border-border"
-      >
-        <input
-          type="text"
-          value={message}
-          onChange={handleMessageChange}
-          placeholder="Type a message according to the topic"
-          className="w-full bg-transparent focus:outline-none"
-        />
-        <button
-          type="submit"
-          className="bg-primary p-2 rounded-lg hover:shadow-hover hover:shadow-primary transition-all duration-300 ease-in-out"
-        >
-          <img src={arrow} alt="submit" />
-        </button>
-      </form>
+      <MessageInput sendMessage={sendMessage} />
     </main>
   );
 }
